end the digest when the last dirty watch is clean
Considering a situation with 100 watches on a scope, when $digested, and only the first of those 100 watches happens to be dirty, that single watch “dirties up” the whole digest round, hence it should run another $digest. On the second round, none of the watches are dirty and the $digest ends with another 100 watch executions. So it had to do 200 watch executions before its done! This can be optimized by caching the last dirty watcher and short circuiting the $digest once the watcher is clean. After optimization the $digest fires only 101 watch executions are done.