continue $digesting on adding new watcher function during the $digest
Adding a watch during the $digest is not considered in the same $digest. The reason is that on the second digest iteration, just before the new watch would run, the $digest ends because it is detecting the first watch as the last dirty watch that’s now clean.
It can be fixed by resetting $$lastDirtyWatch whenever a new watch is added, effectively disabling the optimization.