allow destroying a watcher during digest
Let's say there are three watches. The second watch removes itself when it is called. The expected order would be: [1, 2, 3], [1, 3], since the 2 is removed on first $digest.

But what’s happening instead is that when the second watch removes itself, the watch collection gets shifted to the left, causing $$digestOnce to skip the third watch during that $digest.
So the execution order becomes three cycles; [1, 2], [1, 3], [1, 3]:
 - first for [1, 2] skipping 3, since next forEach call doesn't return anything.
 - second for, [1, 3], second call for 1 and first call for 3.
 - third for [1, 3], second call for 3 and calls 1 because of dirty 3.

If the $$watchers array is reversed, so that new watches are unshifted and iteration is done from the end to the beginning, when a watcher is removed, the part of the watch array that gets shifted has already been handled during that $digest.
So the execution order becomes expected two cycles; [1, 2, 3], [1, 3]:
 - first for [3, 2, 1] in reverse order
  - calls 1
  - removes 2 - shifting 1 to 2's place and the array becomes [3, 1]
  - calls 3 as the next forEachRight will return 3
 - second for [3, 1] in reverse order