evaluate given async function from a watch function even when not dirty
When scheduling async functions in each $digest, on the second time, the watch function won’t be dirty since the value is not changed. That means the $evalAsync also does not run since the $digest has terminated. While it would be run on the next digest, it should really run during the same $digest. So the termination condition now also checks for async queue length and continue digesting until the queue is empty