schedule a digest when evaluating async function
$evalAsync should schedule a $digest if one isn’t already ongoing to insure the async function will be invoked “very soon” instead of waiting for something else to trigger a digest.

When scheduling a $digest, before calling setTimeout, it should check if the queue is empty. If there’s something in the queue, it should have a timeout set and it will eventually drain the queue. Thus scheduling the digest only once. However, inside the setTimeout function it should check the queue is not empty. The queue may have been drained for some other reason before the timeout function was executed, and it is unnecessary to start a $digest with nothing to do.