take array of watch functions and call listener function once for all changes
Starting with simple implementation,

```
watchFns.forEach(watchFn => this.$watch(watchFn, listenerFn));
```

which calls the listener function with individual values instead of array of values. So a custom listener is added to
set an array of values and passed to the watch function:

```
(newValue, oldValue) => {

    newValues[index] = newValue;
    oldValues[index] = oldValue;

    listenerFn(newValues, oldValues, this);

}
```

Now the listener gets called with array of values but for each change. The expected behaviour is that the listener
should be triggered only once for all the changes in the watch group, not for individual changes. So a flag is added to
schedule the listener function to be called later in the digest through $evalAsync:

```
if (!scheduleToTellListener) {

    scheduleToTellListener = true;
    this.$evalAsync(tellListener);

}
```